adapter
Chỉnh sửa adapter cho list viw duyệt thành viên vào nhóm
