+++ app/src/main/res/layout/member_option_layout.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MemberReviewActivity"
+    tools:context=".MemberOptionActivity"
     android:background="@drawable/bg_gradient_main"
     android:padding="15dp">
